refactor(Bonus) : 원시값을 얻는 함수를 이용하도록 리팩토링
