+++ src/main/java/lotto/Application.java
@@ -1,6 +1,5 @@
 package lotto;
 
-import lotto.domain.InputView;
 import lotto.domain.LottoStore;
 
 public class Application {
