+++ src/main/java/lotto/domain/Bonus.java
@@ -1,8 +1,13 @@
 package lotto.domain;
 
 public class Bonus {
-    public static int bonusNumber;
+    private static int bonusNumber;
+
     public Bonus(int bonusNumber) {
         this.bonusNumber = bonusNumber;
+    }
+
+    public static int getBonusNumber() {
+        return bonusNumber;
     }
 }
