+++ src/main/java/lotto/domain/Money.java
@@ -1,6 +1,7 @@
 package lotto.domain;
 
 import lotto.utils.Validation;
+import lotto.views.InputView;
 
 public class Money {
     public static final int LOTTO_PRICE = 1000;
