+++ src/test/java/lotto/ResultTest.java
@@ -1,7 +1,6 @@
 package lotto;
 
 import lotto.domain.Result;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
